Refactoring of the logically coupled types to a shared file
