+++ timesheetService/TimesheetAggregate.js
@@ -1,44 +1,6 @@
 "use strict";
-const SEED_VERSION = 0;
 
-class AggregateBase {
-  constructor(uuid) {
-    if (!uuid)
-      throw new InvalidOperationException("An Id must be supplied and should be a UUID.");
-    this.id = uuid;
-    this.version = SEED_VERSION;
-    this._uncommittedEvents = [];
-    this._routeEvents = {}; //Must be overridden in inherited classes
-  }
-  get aggregateType() {
-    throw "Not Implemented";
-  }
-
-  _applyEvent(record) {
-    let route = this._routeEvents[record.eventsMetadata.eventType];
-    console.log({eventType: record.eventsMetadata.eventType, route: route}); 
-    route(record.event);
-    ++this.version; //Should equal this event's id
-  }
-  _raiseEvent(event) {
-    let timestamp = Date.now(); //milliseconds elapsed since 1 January 1970 00:00:00 UTC
-        
-    let mappedEvent = {
-      "id" : event.id,
-      "timestamp" : timestamp,
-      "eventId" : event.eventId,
-      "eventsMetadata" : {
-        "aggregateType" : this.aggregateType,
-        "eventType" : event.eventType,
-        "timestamp" : timestamp,
-        "sourceLambdaEvent" : JSON.stringify(event.sourceLambdaEvent), // Can not save as a map as it has empty strings in the object, which DyDB is not keen on
-      },
-      "event" : JSON.stringify(event.event) // Can not save as a map as it has empty strings in the object, which DyDB is not keen on
-    };
-    this._applyEvent(mappedEvent);
-    this._uncommittedEvents.push(mappedEvent);
-  }
-}
+var AggregateBase = require("./eventStore").AggregateBase;
 
 class Timesheet extends AggregateBase {
   get aggregateType() {
@@ -46,7 +8,7 @@ class Timesheet extends AggregateBase {
   }
 
   create(createTimesheetPayload) {
-    if (this.version != SEED_VERSION)
+    if (this.version != this.SEED_VERSION)
       throw new InvalidOperationException("Create can only be executed as the first action.");
     let timesheetCreatedEvent = {
       "id" : this.id,
@@ -58,7 +20,7 @@ class Timesheet extends AggregateBase {
     super._raiseEvent(timesheetCreatedEvent);
   }
   update(updateTimesheetPayload) {
-    if (this.version == SEED_VERSION)
+    if (this.version == this.SEED_VERSION)
       throw new InvalidOperationException("Can not update an uninitialised Aggregate, update can not be executed as the first action.");
     if(this.isSubmitted)
       throw new InvalidOperationException("Can not update a submitted timesheet.");
@@ -72,7 +34,7 @@ class Timesheet extends AggregateBase {
     super._raiseEvent(timesheetUpdatedEvent);
   }
   submit(submitTimesheetPayload){
-    if (this.version == SEED_VERSION)
+    if (this.version == this.SEED_VERSION)
       throw new InvalidOperationException("Can not submit an uninitialised Aggregate, submit can not be executed as the first action.");
     if(this.isSubmitted)
       throw new InvalidOperationException("Can not submit a timesheet that is already submitted.");
@@ -105,14 +67,5 @@ class Timesheet extends AggregateBase {
     this.isSubmitted = false;
   }
 }
-class InvalidOperationException extends Error {
-  constructor(message) {
-    super(message);
-    this.name = 'InvalidOperationException';
-  }
-}
-module.exports = {
-  Timesheet:Timesheet,
-  AggregateBase: AggregateBase,
-  InvalidOperationException: InvalidOperationException  
-};
+
+module.exports = Timesheet;
